feat(auth): Correct typo and add JWT authentication
The typo error in the EmployeeFilters was fixed where `fisrt_name` was changed to `first_name`.
A new authentication using JWT (JSON Web Tokens) was added on various files, and new settings for JWT were also appended on settings.py.
A new router is also added on urls.py for the JWT authentication, and an authentication field is activated on the Employee views. A file auth.py was created to handle customized authentication functionalities.
This will establish a secure way to transmit information between parties and prevent unauthorized access.